Exit improvements
