+++ backend/events.js
@@ -1,1 +1,2 @@
+export const SERVER_EXITING = 'server-exiting'
 export const SERVER_RUNNING = 'server-running'
